structural refactor
